+++ peer-assessment.md
@@ -1,0 +1,50 @@
+# Automate
+## [Meet spec] Running the entire analysis requires rendering one Quarto document
+
+Great! I was able to run your paper qmd
+
+## [Meet spec] The analysis runs without errors
+
+Good job! your analysis works. 
+
+## [Meet spec] The analysis produces the expected output
+
+Based on your `paper.qmd` it looks like your analysis produces the expected outcome 
+
+
+## [Meet spec] Data import/cleaning is handled in its own script(s)
+
+ I can see you uploaded your data in its own file and did some cleaning in an R script. looks great.
+ 
+Although, I think you should have your data cleaning in its own folder labeled ``scratch/``. Refer back to Day 1 interactive session afternoon, under instructions. It was a bit confusing finding it I had to click around. Perhaps also label it as `messy_watershed`
+
+# Organize
+## [Meet spec] Raw data is contained in its own folder
+
+Looks good! under your data file I can see that the raw data was stored inside. 
+
+## [Meet spec ] Intermediate outputs are created and saved to a separate folder from raw data
+
+Also great! you have created and output folder and have added your data cleans inside. 
+
+
+## [Meet spec ] At least one piece of functionality has been refactored into a function in its own file
+
+Good!
+
+# Document
+## [Meet spec ] The repo has a README that explains where to find (1) data, (2) analysis script, (3) supporting code, and (4) outputs
+
+The README looks good to me, all you need to do is add a workflow diagram and you will exceed this expectation
+
+## [Not yet] The README includes a flowchart and text explaining how the analysis works
+
+Missing this but overall your almost done Great job!
+
+## [Meet spec] The code is appropriately commented
+
+Your code is well commented and I would just maybe add what Q1 ect mean. 
+
+## [Meet spec ] Variable and function names are descriptive and follow a consistent naming convention
+
+Looks good to me.
